Renames & seg 3 & 5 texture fixes (#19)
* Remove tools/torch

* Auto compile torch and extract

* Re-implement soh's implementation of lus implementation of an array

* Fix static actors in mr

* Update

* Updated torch

* Updated lus

* uncomment

* Fix collision

* Renames

* More renaming

* Collision Renames

* Update collision.c

* Collision Documentation (#640)

* Collision related renames

* Update

* update lus

* Bug fixes

* remove tabs

* fixes

* fixes

* Fix kart textures

* Networking

* net

* Add networking

* Update

* Cleanup

* Fix balloon rendering

* Fix lakitu and red shell

* fixes

* Implement new random system

---------

+++ src/code_80057C60.h
@@ -380,7 +380,7 @@ extern s16 D_80165900;
 extern s8  D_80165908;
 
 extern s8 D_80165A90;
-// extern Objects gObjectList[]; -> objects.h
+// extern Object gObjectList[]; -> objects.h
 
 extern s32 objectListSize;
 extern Mtx D_80183D60;
@@ -542,7 +542,7 @@ extern u8  D_8018D228;
 extern s32 D_8018D22C;
 // extern s8 D_8018D230; -> objects.h
 
-extern s32 D_8018D240;
+extern uintptr_t D_8018D240;
 extern u8 *D_8018D248[];
 extern f32 D_8018D2A0;
 extern s32 D_8018D2A4;
